Added encoder buffer library
Added the base library files (two header files, and one source file) complete with very basic doxygen documentation and the corresponding html/LaTeX output files (see html/index.html, majority of documentation located under "classes"). Also included is a documented example file to show basic functionality, with pin maps for both the MEGA and Due, as well as a .jpg image for help with the Due SPI hookups. The code has been tested on both the MEGA and the Due  (SPI clock frequencies up to 8MHz on the Mega and 4MHz on the Due). This library still needs to be tested with a working encoder.